+++ sdk/lib/io/common.dart
@@ -15,6 +15,21 @@ const int _errorResponseErrorType = 0;
 const int _osErrorResponseErrorCode = 1;
 const int _osErrorResponseMessage = 2;
 
+// POSIX error codes.
+// See https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/errno.h.html
+const _eNoEnt = 2;
+
+// Windows error codes.
+// See https://learn.microsoft.com/en-us/windows/win32/debug/system-error-codes--0-499-
+const _errorFileNotFound = 2;
+const _errorPathNotFound = 3;
+const _errorInvalidDrive = 15;
+const _errorNoMoreFiles = 18;
+const _errorBadNetpath = 53;
+const _errorBadNetName = 67;
+const _errorBadPathName = 161;
+const _errorFilenameExedRange = 206;
+
 /// If the [response] is an error, throws an [Exception] or an [Error].
 void _checkForErrorResponse(Object? response, String message, String path) {
   if (response is List<Object?> && response[0] != _successResponse) {
@@ -24,7 +39,7 @@ void _checkForErrorResponse(Object? response, String message, String path) {
       case _osErrorResponse:
         var err = OSError(response[_osErrorResponseMessage] as String,
             response[_osErrorResponseErrorCode] as int);
-        throw FileSystemException(message, path, err);
+        throw FileSystemException._fromOSError(err, message, path);
       case _fileClosedResponse:
         throw FileSystemException("File closed", path);
       default:
@@ -64,12 +79,18 @@ class OSError implements Exception {
     StringBuffer sb = new StringBuffer();
     sb.write("OS Error");
     if (message.isNotEmpty) {
-      sb..write(": ")..write(message);
+      sb
+        ..write(": ")
+        ..write(message);
       if (errorCode != noErrorCode) {
-        sb..write(", errno = ")..write(errorCode.toString());
+        sb
+          ..write(", errno = ")
+          ..write(errorCode.toString());
       }
     } else if (errorCode != noErrorCode) {
-      sb..write(": errno = ")..write(errorCode.toString());
+      sb
+        ..write(": errno = ")
+        ..write(errorCode.toString());
     }
     return sb.toString();
   }
@@ -85,8 +106,7 @@ class _BufferAndStart {
 // Ensure that the input List can be serialized through a native port.
 _BufferAndStart _ensureFastAndSerializableByteData(
     List<int> buffer, int start, int end) {
-  if ((buffer is Uint8List) &&
-      (buffer.buffer.lengthInBytes == buffer.length)) {
+  if ((buffer is Uint8List) && (buffer.buffer.lengthInBytes == buffer.length)) {
     // Send typed data directly, unless it is a partial view, in which case we
     // would rather copy than drag in the potentially much large backing store.
     // See issue 50206.